Withdraws and separates database adapters. Split local and remote database tasks.
