+++ lib/capistrano-db-tasks.rb
@@ -1,2 +1,2 @@
-$stderr.puts %Q(Please, change \"require 'capistrano-db-tasks'\" to \"require 'capistrano/db_tasks'" in your Capfile.)
+$stderr.puts %(Please, change \"require 'capistrano-db-tasks'\" to \"require 'capistrano/db_tasks'" in your Capfile.)
 require 'capistrano/db_tasks'
