+++ lib/capistrano/db_tasks.rb
@@ -6,4 +6,12 @@ require_relative 'db_tasks/compressors/base'
 require_relative 'db_tasks/compressors/bzip2'
 require_relative 'db_tasks/compressors/gzip'
 
+require_relative 'db_tasks/databases/base'
+require_relative 'db_tasks/databases/local'
+require_relative 'db_tasks/databases/remote'
+
+require_relative 'db_tasks/adapters/base'
+require_relative 'db_tasks/adapters/mysql'
+require_relative 'db_tasks/adapters/postgres'
+
 require_relative 'db_tasks/tasks'
